requirement.txt
amqp==2.4.2
asn1crypto==0.24.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.4.16
cffi==1.12.2
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.6.1
decorator==4.3.0
Django==1.11.13
django-cors-headers==2.5.1
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
httpie==0.9.9
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
kombu==4.4.0
MarkupSafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.4
redis==3.2.0
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.2.0
wcwidth==0.1.7
